tools/requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
idna==2.9
importlib-metadata==1.6.0
jsonpickle==1.4.1
python-digitalocean==1.15.0
requests==2.23.0
urllib3==1.25.9
zipp==3.1.0
